+++ src/components/PeopleTable.tsx
@@ -1,0 +1,56 @@
+import { useParams } from 'react-router-dom';
+import { Person } from '../types';
+import PersonLink from './PersonLink';
+
+type Props = {
+  people: Person[];
+};
+
+export default function PeopleTable({ people }: Props) {
+  const { slug } = useParams();
+
+  return (
+    <table
+      data-cy="peopleTable"
+      className="table is-striped is-hoverable is-fullwidth"
+    >
+      <thead>
+        <tr>
+          <th>Name</th>
+          <th>Sex</th>
+          <th>Born</th>
+          <th>Died</th>
+          <th>Mother</th>
+          <th>Father</th>
+        </tr>
+      </thead>
+      <tbody>
+        {people.map(person => {
+          const mother = people.find(p => p.name === person.motherName);
+          const father = people.find(p => p.name === person.fatherName);
+
+          return (
+            <tr
+              data-cy="person"
+              key={person.slug}
+              className={slug === person.slug ? 'has-background-warning' : ''}
+            >
+              <td>
+                <PersonLink person={person} name={person.name} />
+              </td>
+              <td>{person.sex}</td>
+              <td>{person.born}</td>
+              <td>{person.died}</td>
+              <td>
+                <PersonLink person={mother} name={person.motherName} />
+              </td>
+              <td>
+                <PersonLink person={father} name={person.fatherName} />
+              </td>
+            </tr>
+          );
+        })}
+      </tbody>
+    </table>
+  );
+}
